+++ src/LocoMotor/UITextLM.h
@@ -36,8 +36,8 @@ namespace LocoMotor {
 		void AlignLeft();
 		void AlignRight();
 
-		void SetTopColor(int r, int g, int b);
-		void SetBottomColor(int r, int g, int b);
+		void SetTopColor(double r, double g, double b);
+		void SetBottomColor(double r, double g, double b);
 
 	protected:
 		OgreWrapper::UIText* _uTxt;
